// src/common/enums/notifications.ts
export enum Source {
  CLIENT = 'CLIENT',
  BRANCH = 'BRANCH',
  SERVICE = 'SERVICE',
  AWARENESS_CAMP = 'AWARENESS_CAMP',
  CUSTOMER_CARE = 'CUSTOMER_CARE',
}

export enum NOTIFICATION_TYPES {
  SERVICE_PLANNED = 'SERVICE_PLANNED',
  SERVICE_COMPLETED = 'SERVICE_COMPLETED',
  SERVICE_CANCELLED = 'SERVICE_CANCELLED',
  SERVICE_RESCHEDULED = 'SERVICE_RESCHEDULED',
  SERVICE_ACTIVATED = 'SERVICE_ACTIVATED',
  SERVICE_DEACTIVATED = 'SERVICE_DEACTIVATED',
  SERVICE_DELETED = 'SERVICE_DELETED',
  SERVICE_STATUS_UPDATED = 'SERVICE_STATUS_UPDATED',
  TICKET_RAISED = 'TICKET_RAISED',
  TICKET_UPDATED = 'TICKET_UPDATED',
  TICKET_COMPLETED = 'TICKET_COMPLETED',
  TICKET_CANCELLED = 'TICKET_CANCELLED',
  AWARENESSCAMP_PLANNED = 'AWARENESSCAMP_PLANNED',
  AWARENESSCAMP_COMPLETED = 'AWARENESSCAMP_COMPLETED',
  AWARENESSCAMP_REJECT = 'AWARENESSCAMP_REJECT',
  AWARENESSCAMP_UPDATED = 'AWARENESSCAMP_UPDATED',
  AWARENESSCAMP_DELETED = 'AWARENESSCAMP_DELETED',
}

export const notificationStatusColor = {
  'On going': '#8E4B10',
  Ongoing: '#8E4B10',
  Planned: '#1A56DB',
  Rescheduled: '#1A56DB',
  Completed: '#046C4E',
  Cancelled: '#AB0106',
};